fhdev/requirements.txt
Django==2.0.6
django-cors-headers==2.4.0
django-summernote==0.8.10.0
django-templated-mail==1.1.1
djangorestframework==3.9.0
djangorestframework-jsonapi==2.6.0
djangorestframework-jwt==1.11.0
djoser==1.3.1
inflection==0.3.1
PyJWT==1.7.0
pytz==2018.7
six==1.11.0
